Applied Accessibility
In web development, accessibility refers to web content and a UI (user interface) that can be understood, navigated, and interacted with by a broad audience. This includes people with visual, auditory, mobility, or cognitive disabilities.

In this course, you'll learn best practices for building webpages that are accessible to everyone.